200726 | added posts Meta OG api

+++ src/main/java/com/hyunro/bookmark/service/jwt/JwtService.java
@@ -5,7 +5,7 @@ import java.util.Map;
 public interface JwtService {
     <T> String create(String key, T data, String subject);
     Map<String, Object> get(String key);
-    int getUserId();
+    Long getUserId();
     boolean isUsable(String jwt);
 
 }
